Add files via upload
In this project, we have developed a predictive model for the Health Insurance Marketplace in the US. We examined the transitions between 2014, 2015 and 2016 before and after the Affordable Care Act’s coverage-related provisions took effect in 2014. Although, close to 22.8 million people gained insurance coverage after the act, the ACA came out successful, however, 5.9 million people lost their coverage, for a net increase of 16.9 million people with insurance. With in-depth research, trend and exploratory analysis of the data released by the ‘Centers for Medicare & Medicaid Services (CMS)’ with over twelve million records, we found significant hike in the Maximum Out of Pocket (MOOP) costs that the families bear after the issuer pays the rest. Considering the various state ideologies in the country, we found that with increasing liberal score, the monthly premium rate is significantly lower than the conservative states. We also developed a classification models using C50 and rpart algorithms and evaluated the models on the test sets to evaluate their acccuracies. Subsequently, we successfully trained a neural network model on the data for years 2014-2016 to predict the plan rates for the year 2017.